PCI: microchip: Move setup functions to pcie-plda-host.c
Move setup functions to common pcie-plda-host.c. So these two functions
can be re-used.
